requirements.txt
aiohttp==3.9.3
APScheduler==3.10.4
motor
pymongo==4.8.0
git+https://github.com/KurimuzonAkuma/pyrogram
TgCrypto==1.2.3
pyromod==3.1.6
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
uvloop==0.19.0
dextools_python